refactor: move node icon definitions to web config
